+++ sbt-scalajs-crossproject/src/main/scala/org/scalajs/sbtplugin/JSPlatform.scala
@@ -1,10 +1,11 @@
-package scalajscrossproject
+/* It is totally evil to introduce a class in the package of sbt-scalajs, but that is
+ * the only way we can be forward source compatible with Scala.js 1.x.
+ */
+package org.scalajs.sbtplugin
 
 import sbtcrossproject._
 
 import sbt._
-
-import org.scalajs.sbtplugin.{ScalaJSCrossVersion, ScalaJSPlugin}
 
 case object JSPlatform extends Platform {
   def identifier: String                = "js"
